[Update] checkLogin and checkCreator functions are separated from main to separate Blueprint
